pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.128.0",
    "gymnasium>=1.2.3",
    "httpx>=0.28.1",
    "langchain>=1.2.3",
    "langchain-openai>=1.1.7",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.6",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "plotly>=6.5.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.0",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.45",
    "stable-baselines3>=2.7.1",
    "tensorflow>=2.20.0",
    "uvicorn>=0.40.0",
    "yfinance>=1.0",
    # Development tools
    "black>=24.8.0",
    "flake8>=7.1.1",
    "isort>=5.13.2",
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
]
[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"



